Update module namespace to Adobe
