+++ Adobe/AxpConnector/etc/adminhtml/system.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Config:etc/system_file.xsd">
+    <system>
+        <tab id="axpconnector" translate="label" sortOrder="2000">
+            <label>Adobe Experience Platform</label>
+        </tab>
+        <section id="axpconnector_backend_config" translate="label" type="text" sortOrder="0" showInDefault="1" showInWebsite="1" showInStore="1">
+            <label>Configuration</label>
+            <tab>axpconnector</tab>
+            <resource>Adobe_AxpConnector::axpconnector_config</resource>
+            <class>separator-top</class>
+            <group id="general" translate="label" sortOrder="2" showInDefault="1" showInWebsite="0" showInStore="0">
+                <field id="enable" translate="label" type="select" sortOrder="100" showInDefault="1" showInWebsite="0" showInStore="0">
+                    <label>Module Enable</label>
+                    <source_model>Magento\Config\Model\Config\Source\Yesno</source_model>
+                </field>
+                <field id="launch_script_url" translate="label" type="text" sortOrder="200" showInDefault="1" showInWebsite="0" showInStore="0">
+                    <label>Launch Script URL</label>
+                    <comment>The script URL (or script tag) from the Launch interface.</comment>
+                    <backend_model>Adobe\AxpConnector\Model\Config\Backend\LaunchScriptUrl</backend_model>
+                </field>
+            </group>
+        </section>
+    </system>
+</config>
