+++ Adobe/AxpConnector/etc/frontend/events.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
+    <event name="checkout_cart_add_product_complete">
+        <observer name="Adobe_axpconnector_add_to_cart" instance="Adobe\AxpConnector\Observer\CheckoutCartAddProductObserver" />
+    </event>
+    <event name="sales_quote_remove_item">
+        <observer name="Adobe_axpconnector_remove_from_cart" instance="Adobe\AxpConnector\Observer\SalesQuoteRemoveItemObserver" />
+    </event>
+    <event name="checkout_onepage_controller_success_action">
+        <observer name="Adobe_axpconnector_order_success" instance="Adobe\AxpConnector\Observer\CheckoutOnepageControllerSuccessActionObserver" />
+    </event>
+    <event name="multishipping_checkout_controller_success_action">
+        <observer name="Adobe_axpconnector_order_success" instance="Adobe\AxpConnector\Observer\CheckoutOnepageControllerSuccessActionObserver" />
+    </event>
+</config>
