+++ Adobe/AxpConnector/registration.php
@@ -1,0 +1,6 @@
+<?php
+\Magento\Framework\Component\ComponentRegistrar::register(
+  \Magento\Framework\Component\ComponentRegistrar::MODULE,
+  'Adobe_AxpConnector',
+  __DIR__
+);
