+++ Adobe/AxpConnector/view/frontend/layout/checkout_cart_index.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <body>
+        <referenceContainer name="after.body.start">
+            <block class="Adobe\AxpConnector\Block\Cart" name="axpconnector_cart_block" template="Adobe_AxpConnector::cart.phtml" after="-" />
+        </referenceContainer>
+    </body>
+</page>
