+++ Adobe/AxpConnector/view/frontend/layout/default.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0"?>
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">    
+    <body>
+        <referenceBlock name="head.additional">
+            <block name="axpconnector_launch_head_script" before="-" template="Adobe_AxpConnector::launch_head.phtml" />
+        </referenceBlock>
+
+        <referenceContainer name="after.body.start">
+            <block class="Adobe\AxpConnector\Block\Display" name="axpconnector_datalayer_block" as="axpconnector_datalayer_block" template="Adobe_AxpConnector::datalayer.phtml">
+            </block>
+
+            <block name="axpconnector_private_content_block"  template="Adobe_AxpConnector::private_content_datalayer.phtml">
+              <arguments>
+                <argument name="jsLayout" xsi:type="array">
+                  <item name="components" xsi:type="array">
+                    <item name="launch" xsi:type="array">
+                      <item name="component" xsi:type="string">Adobe_AxpConnector/js/view/launch</item>
+                    </item>
+                  </item>
+                </argument>
+              </arguments>
+            </block>
+
+            <block class="Adobe\AxpConnector\Block\Page" name="axpconnector_page_block" template="Adobe_AxpConnector::page.phtml" />
+        </referenceContainer>
+    </body>
+</page>
