+++ Adobe/AxpConnector/view/frontend/requirejs-config.js
@@ -1,0 +1,7 @@
+var config = {
+    map: {
+        '*': {
+            axpconnector: 'Adobe_AxpConnector/js/axpconnector'
+        }
+    }
+};
