+++ Adobe/AxpConnector/view/frontend/templates/launch_head.phtml
@@ -1,0 +1,7 @@
+<?php
+$_helper = $this->helper('Adobe\AxpConnector\Helper\Data');
+?>
+
+<?php if ($_helper->isEnabled()) : ?>
+  <script src="<?php echo $_helper->getScriptUrl(); ?>" async></script>
+<?php endif; ?>
