+++ Adobe/AxpConnector/view/frontend/templates/product.phtml
