+++ Adobe/AxpConnector/view/frontend/web/js/view/launch.js
