--- SearchDiscovery/AxpConnector/composer.json
@@ -1,21 +1,0 @@
-{
-    "name": "searchdiscovery/module-axpconnector",
-    "description": "Search Discovery Launch By Adobe Module",
-    "require": {
-        "php": "~5.6.5|7.0.2|7.0.4|~7.0.6|~7.1.0"
-    },
-    "type": "magento2-module",
-    "version": "0.1.0",
-    "license": [
-        "OSL-3.0",
-        "AFL-3.0"
-    ],   
-    "autoload": {
-        "files": [
-            "registration.php"
-        ],
-        "psr-4": {
-            "SearchDiscovery\\AxpConnector\\": ""
-        }
-    }
-}
