--- SearchDiscovery/AxpConnector/etc/frontend/di.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <type name="Magento\Customer\CustomerData\SectionPoolInterface">
-        <arguments>
-            <argument name="sectionSourceMap" xsi:type="array">
-                <item name="launch" xsi:type="string">SearchDiscovery\AxpConnector\CustomerData\Launch</item>
-            </argument>
-        </arguments>
-    </type>
-</config>
