--- SearchDiscovery/AxpConnector/etc/frontend/events.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
-    <event name="checkout_cart_add_product_complete">
-        <observer name="searchdiscovery_axpconnector_add_to_cart" instance="SearchDiscovery\AxpConnector\Observer\CheckoutCartAddProductObserver" />
-    </event>
-    <event name="sales_quote_remove_item">
-        <observer name="searchdiscovery_axpconnector_remove_from_cart" instance="SearchDiscovery\AxpConnector\Observer\SalesQuoteRemoveItemObserver" />
-    </event>
-    <event name="checkout_onepage_controller_success_action">
-        <observer name="searchdiscovery_axpconnector_order_success" instance="SearchDiscovery\AxpConnector\Observer\CheckoutOnepageControllerSuccessActionObserver" />
-    </event>
-    <event name="multishipping_checkout_controller_success_action">
-        <observer name="searchdiscovery_axpconnector_order_success" instance="SearchDiscovery\AxpConnector\Observer\CheckoutOnepageControllerSuccessActionObserver" />
-    </event>
-</config>
