--- SearchDiscovery/AxpConnector/view/frontend/layout/checkout_index_index.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceContainer name="after.body.start">
-            <block class="SearchDiscovery\AxpConnector\Block\Checkout" name="axpconnector_checkout_block" template="SearchDiscovery_AxpConnector::checkout.phtml" after="-" />
-        </referenceContainer>
-    </body>
-</page>
