--- SearchDiscovery/AxpConnector/view/frontend/requirejs-config.js
@@ -1,7 +1,0 @@
-var config = {
-    map: {
-        '*': {
-            axpconnector: 'SearchDiscovery_AxpConnector/js/axpconnector'
-        }
-    }
-};
