--- SearchDiscovery/AxpConnector/view/frontend/templates/launch_head.phtml
@@ -1,7 +1,0 @@
-<?php
-$_helper = $this->helper('SearchDiscovery\AxpConnector\Helper\Data');
-?>
-
-<?php if ($_helper->isEnabled()) : ?>
-  <script src="<?php echo $_helper->getScriptUrl(); ?>" async></script>
-<?php endif; ?>
